+++ Updatum.FakeApp/Program.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using StageKit.Runtime;
+using System;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Threading.Tasks;